refactor : 예약하기 버튼 분리
checkExist 함수는 예약상품 목록 변경될때마다 새로 생성됨. TravelItemCard 에서 해당 함수 구독하면 불필요한 리렌더링 발생.
